% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_launch.R
\name{launch_tutorial}
\alias{launch_tutorial}
\title{launch_tutorial function}
\usage{
launch_tutorial(test = FALSE)
}
\arguments{
\item{test}{A logical for developing function in inst/tutorials dir vs. installing function in tutorials dir}
}
\description{
This function allows you to launch the bespoke tutorial you previously made
}
\keyword{tutorial}
