% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{make_lecture}
\alias{make_lecture}
\title{Make Lecture Function}
\usage{
make_lecture(list, output_file)
}
\arguments{
\item{list}{Numeric vector of numbers corresponding to the essential lectures you want to use}
}
\description{
This function allows you to make bespoke lectures
}
\examples{
make_lecture(list = c(0, 7, 10))
}
\keyword{lecture}
