% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mtcars}
\alias{mtcars}
\title{mtcars}
\format{
A data frame with 20,430 rows and 8 variables:
\describe{
\item{id}{Make and model}
\item{make}{Make of the car}
\item{model}{Model of the car}
\item{mpg}{Miles/(US) gallon}
\item{cyl}{Number of cylinders}
\item{disp}{Displacement (cu.in.)}
\item{hp}{Gross horsepower}
\item{drat}{Rear axle ratio}
\item{wt}{Weight (1000 lbs)}
\item{qsec}{1/4 mile time}
\item{vs}{Engine (0 = V-shaped, 1 = straight)}
\item{am}{Transmission (0 = automatic, 1 = manual)}
\item{gear}{Number of forward gears}
\item{carb}{Number of carburetors}
}
}
\source{
{Henderson and Velleman (1981), Building multiple regression models interactively. Biometrics, 37, 391–411}
}
\usage{
data(mtcars)
}
\description{
A dataset containing human proteome information from Uniprot.
}
\keyword{datasets}
