% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_prepare.R
\name{prepare_data}
\alias{prepare_data}
\title{prepare_data function}
\usage{
prepare_data(bespoke_dataframe = NULL, test = FALSE)
}
\arguments{
\item{bespoke_dataframe}{name of an object in your global environment from which to create a bespoke dataset}

\item{test}{A logical for developing function in inst/tutorials dir vs. installing function in tutorials dir}
}
\description{
This function allows you to prepare a data set for bespoke content generation
}
\keyword{prepare}
