% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_prepare.R
\name{prepare_metadata}
\alias{prepare_metadata}
\title{prepare_metadata Function}
\usage{
prepare_metadata(
  bespoke_dataframe = NULL,
  dataframe_name = "INSERT_NAME",
  dataframe_title = "INSERT_TITLE",
  dataframe_source = "INSERT_SOURCE",
  dataframe_desc = "INSERT_DESC",
  dataframe_about = "INSERT_ABOUT",
  dataframe_join_name = "INSERT_JOIN_NAME",
  dataframe_join_about = "INSERT_JOIN_ABOUT",
  test = FALSE
)
}
\arguments{
\item{bespoke_dataframe}{name of an object in your global environment from which you created a bespoke dataset}

\item{dataframe_name}{Name of your dataframe}

\item{dataframe_title}{Title of your dataframe}

\item{dataframe_source}{Source of your dataframe}

\item{dataframe_desc}{Description of your dataframe}

\item{dataframe_about}{About your dataframe}

\item{dataframe_join_name}{Name of your dataframe for joining}

\item{dataframe_join_about}{About your dataframe for joining}

\item{test}{A logical for developing function in inst/tutorials dir vs. installing function in tutorials dir}
}
\description{
This function allows you to prepare metadata for your bespoke dataset. If from_template = FALSE, then the function is expecting a bunch of variables to be assigned. If from_template = TRUE, then a 'dummy' template is made. This step is not necessary if using default mtcars dataset
}
\keyword{metadata}
