% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_prepare.R
\name{prepare_yaml}
\alias{prepare_yaml}
\title{prepare_yaml function}
\usage{
prepare_yaml(
  test = FALSE,
  title = "INSERT_TITLE",
  subtitle = "INSERT_SUBTITLE",
  author = "INSERT_AUTH",
  date = lubridate::today()
)
}
\arguments{
\item{test}{A logical for developing function in inst/tutorials dir vs. installing function in tutorials dir}

\item{bespoke_dataframe}{name of an object in your global environment from which to create a bespoke dataset}
}
\description{
This function allows you to prepare the YAML header for bespoke content generation
}
\keyword{prepare}
