% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_launch.R
\name{launch_lecture}
\alias{launch_lecture}
\title{launch_lecture function}
\usage{
launch_lecture()
}
\description{
This function allows you to launch the bespoke lecture you previously made
}
\examples{
launch_lecture()
}
\keyword{tutorial}
