% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_make.R
\name{make_lecture}
\alias{make_lecture}
\title{Make Lecture Function}
\usage{
make_lecture(lecture_num, output_file)
}
\arguments{
\item{lecture_num}{Numeric vector of numbers corresponding to the essential lectures you want to use}

\item{output_file}{Path to place compiled file in 'tutorials/bespoketutorial' dir}
}
\description{
This function allows you to make bespoke lectures
}
\examples{
make_lecture(list = c(0, 7, 10))
}
\keyword{lecture}
