% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_make.R
\name{make_list}
\alias{make_list}
\title{make_list function}
\usage{
make_list(lecture_num)
}
\arguments{
\item{list}{Numeric vector of numbers corresponding to the essential lectures you want to use}
}
\description{
This internal function allows you to make a list of lectures for bespoke content generation
}
