% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_make.R
\name{make_tutorial}
\alias{make_tutorial}
\title{make_tutorial Function}
\usage{
make_tutorial(lecture_num, test = FALSE)
}
\arguments{
\item{lecture_num}{Numeric vector of numbers corresponding to the essential lectures you want to use}

\item{test}{A logical for developing function in inst/tutorials dir vs. installing function in tutorials dir}
}
\description{
This function allows you to make bespoke tutorials
}
\keyword{tutorial}
